Rename "components" in playground to "app"
