+++ packages/playground/package.json
@@ -7,8 +7,7 @@
     "monaco-editor-webpack-plugin": "^4.2.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-monaco-editor": "^0.45.0",
-    "tailwindcss-classnames": "^2.2.5"
+    "react-monaco-editor": "^0.45.0"
   },
   "scripts": {
     "build": "rimraf dist pkg && webpack --config webpack.config.js --mode production",
