--- packages/playground/src/ids.ts
@@ -1,1 +1,0 @@
-export const buttonIcon = 'buttonIcon';
